rewrote script.js to work with paramters
