+++ Dockerfile
@@ -1,10 +1,15 @@
-FROM ubuntu
+FROM ghcr.io/puppeteer/puppeteer:latest
 
-WORKDIR myDir
+# root user to be able to access ount directory
+USER root
+
+WORKDIR supercias
 # pass our own package.json
 COPY package.json package.json
 COPY options.json options.json
-# install files
-#RUN npm install 
+COPY src/ src/
+# install files 
+RUN npm install 
 
+# works with this command
 #docker run -it --init --cap-add=SYS_ADMIN -v /home/telix/supercias/data/:/home/pptruser/supercias/data supercias_image npm run companies
