--- Dockerfile.bkp
@@ -1,10 +1,0 @@
-FROM ghcr.io/puppeteer/puppeteer:latest
-
-# WORKDIR ../supercias
-# pass our own package.json
-COPY package.json package.json
-COPY options.json options.json
-# install files
-RUN npm install 
-
-#docker run -it --init --cap-add=SYS_ADMIN -v /home/telix/supercias/data/:/home/pptruser/supercias/data supercias_image npm run companies
