+++ src/utils/getText.js
