+++ src/utils/imprint2Function.js
