+++ src/utils/strToBinary.js
